Fix edit layer component (#703)
* fixed fl-simpledatetime

* invalid date

* test

* Доработка результатов идентификации

---------
