+++ app/helpers/format-property.js
@@ -1,0 +1,1 @@
+export { default, formatProperty } from 'ember-flexberry-gis/helpers/format-property';
